Update conda recipe to build graphblas-opt
This commit makes the conda recipe build graphblas-opt.

We had to add some temporary changes to make it so that we use the dev label from the metagraph channel for the mlir conda package. This necessitated some changes in the tests.

NOTE: graphblas-opt so far is simply a copy-and-rename of the standalone template from https://github.com/llvm/llvm-project/tree/main/mlir/examples/standalone

For posterity, below is a flimsy script used to copy-and-rename the template:

        # don't use these exports
        # export PREFIX=/Users/pnguyen/anaconda3/envs/mlirgraphblas
        # export BUILD_DIR=/Users/pnguyen/anaconda3/envs/mlirgraphblas
        # export LD_LIBRARY_PATH=/Users/pnguyen/anaconda3/envs/mlirgraphblas/lib
        # export PYTHONPATH=/Users/pnguyen/anaconda3/envs/mlirgraphblas/lib/python3.8/site-packages

        # using llvm-project commit e413b86
        cd ~/code/mlir-graphblas/mlir_graphblas
        rm -rf src
        cp -r /tmp/test/llvm-project/mlir/examples/standalone/ ./src/
        git checkout src/build.py
        cd src
        rm -rf README.md
        black $(find . -name "*py")

        # Rename files
        rm -rf ./standalone-translate
        rm -rf ./test/Standalone/standalone-translate.mlir
        mv ./test/Standalone ./test/GraphBLAS
        mv ./test/GraphBLAS/standalone-opt.mlir ./test/GraphBLAS/graphblas-opt.mlir
        mv ./include/Standalone ./include/GraphBLAS
        mv ./include/GraphBLAS/StandaloneOps.td ./include/GraphBLAS/GraphBLASOps.td
        mv ./include/GraphBLAS/StandaloneOps.h ./include/GraphBLAS/GraphBLASOps.h
        mv ./include/GraphBLAS/StandaloneDialect.h ./include/GraphBLAS/GraphBLASDialect.h
        mv ./include/GraphBLAS/StandaloneDialect.td ./include/GraphBLAS/GraphBLASDialect.td
        mv ./lib/Standalone ./lib/GraphBLAS
        mv ./lib/GraphBLAS/StandaloneDialect.cpp ./lib/GraphBLAS/GraphBLASDialect.cpp
        mv ./lib/GraphBLAS/StandaloneOps.cpp ./lib/GraphBLAS/GraphBLASOps.cpp
        mv ./standalone-opt ./graphblas-opt
        mv ./graphblas-opt/standalone-opt.cpp ./graphblas-opt/graphblas-opt.cpp
        for e in $(grep -irl "standalone" .) ; do sed -i'~' 's/STANDALONE/GRAPHBLAS/g' $e ; done ; rm -rf $(find . -name "*~")
        for e in $(grep -irl "standalone" .) ; do sed -i'~' 's/Standalone/GraphBLAS/g' $e ; done ; rm -rf $(find . -name "*~")
        for e in $(grep -irl "standalone" .) ; do sed -i'~' 's/standalone/graphblas/g' $e ; done ; rm -rf $(find . -name "*~")
        for e in $(grep -irl "add_subdirectory.graphblas-translate." .) ; do sed -i'~' 's/add_subdirectory(graphblas-translate)//g' $e ; done ; rm -rf $(find . -name "*~")
        for e in $(grep -irl "        graphblas-translate" .) ; do sed -i'~' 's/        graphblas-translate//g' $e ; done ; rm -rf $(find . -name "*~")
        for e in $(grep -irl "tools = \[\"graphblas-opt\", \"graphblas-translate\"\]" .) ; do sed -i'~' 's/, "graphblas-translate"//g' $e ; done ; rm -rf $(find . -name "*~")
        python3 build.py

        # Ignore these
        # mkdir build
        # cd build
        # cmake -G Ninja .. -DMLIR_DIR=$PREFIX/lib/cmake/mlir -DLLVM_EXTERNAL_LIT=$BUILD_DIR/bin/llvm-lit
        # cmake --build . --target check-graphblas

        # We must manually add "MLIRIR" to be a member of LIBS in mlir-graphblas/mlir_graphblas/src/graphblas-opt/CMakeLists.txt
        # We got the solution from the last commit linked in https://github.com/tensorflow/mlir/issues/289

+++ mlir_graphblas/src/build.py
@@ -62,14 +62,14 @@ if __name__ == "__main__":
         "-build-clean", action="store_true", help="Rebuild from scratch."
     )
     args = parser.parse_args()
-    if args.build_clean:
+    if args.build_clean and os.path.isdir(build_dir):
         shutil.rmtree(build_dir)
 
     if not os.path.isdir(build_dir):
         os.makedirs(build_dir)
 
     env_lib_path = os.path.join(sys.exec_prefix, "lib")
-    LD_LIBRARY_PATH = os.environ.get("LD_LIBRARY_PATH", "")
+    LD_LIBRARY_PATH = os.environ.get("LD_LIBRARY_PATH", env_lib_path)
     if env_lib_path not in LD_LIBRARY_PATH.split(":"):
         LD_LIBRARY_PATH = LD_LIBRARY_PATH + ":" + env_lib_path
 
@@ -78,7 +78,7 @@ if __name__ == "__main__":
     command, stdout_string, stderr_string = run_shell_commands(
         build_dir,
         "cmake -G Ninja .. -DMLIR_DIR=$PREFIX/lib/cmake/mlir -DLLVM_EXTERNAL_LIT=$BUILD_DIR/bin/llvm-lit",  # creates ./build/graphblas-opt
-        "cmake --build . --target check-graphblas",  # creates ./build/bin/graphblas-opt and runs tests
+        "cmake --build . --target check-graphblas --verbose",  # creates ./build/bin/graphblas-opt and runs tests
         PREFIX=sys.exec_prefix,
         BUILD_DIR=sys.exec_prefix,
         LD_LIBRARY_PATH=LD_LIBRARY_PATH,
